Combined ProcessInstanceQuery and TaskQuery clients into SearchService.
