+++ kie-server-parent/kie-server-remote/kie-server-client/src/main/java/org/kie/server/client/impl/SearchServicesClientImpl.java
@@ -18,36 +18,48 @@ package org.kie.server.client.impl;
 import static org.kie.server.api.rest.RestURI.*;
 
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
+import org.kie.server.api.model.instance.ProcessInstance;
+import org.kie.server.api.model.instance.ProcessInstanceList;
 import org.kie.server.api.model.instance.TaskInstance;
 import org.kie.server.api.model.instance.TaskInstanceList;
 import org.kie.server.client.KieServicesConfiguration;
-import org.kie.server.client.TaskQueryServicesClient;
+import org.kie.server.client.SearchServicesClient;
+import org.kie.server.jbpm.queries.api.model.definition.ProcessInstanceQueryFilterSpec;
 import org.kie.server.jbpm.queries.api.model.definition.TaskQueryFilterSpec;
 
-public class TaskQueryServicesClientImpl extends AbstractKieServicesClientImpl implements TaskQueryServicesClient {
+public class SearchServicesClientImpl extends AbstractKieServicesClientImpl implements SearchServicesClient {
 
-	private static Set<Class<?>> extraClasses;
-
-	static {
-		extraClasses = new HashSet<>();
-		extraClasses.add(TaskQueryFilterSpec.class);
-	}
-
-	// TODO: Adding extra classes to the config at this point doesn't seem to work. They're not picked up by the JaxB Marshaller
-	public TaskQueryServicesClientImpl(KieServicesConfiguration config) {
+	public SearchServicesClientImpl(KieServicesConfiguration config) {
 		super(config);
-		config.addExtraClasses(extraClasses);
 	}
-
-	public TaskQueryServicesClientImpl(KieServicesConfiguration config, ClassLoader classLoader) {
+	
+	public SearchServicesClientImpl(KieServicesConfiguration config, ClassLoader classLoader) {
 		super(config, classLoader);
-		config.addExtraClasses(extraClasses);
 	}
 
+	@Override
+	public List<ProcessInstance> findProcessInstancesWithFilters(ProcessInstanceQueryFilterSpec filterSpec, Integer page, Integer pageSize) {
+		ProcessInstanceList result = null;
+
+		if (config.isRest()) {
+			String queryString = getPagingQueryString("?", page, pageSize);
+			result = makeHttpPostRequestAndCreateCustomResponse(loadBalancer.getUrl() + "/" + PROCESS_INSTANCES_GET_FILTERED_URI + queryString,
+					filterSpec, ProcessInstanceList.class);
+
+		} else {
+			// TODO: Need to implement the command (used for non-REST scenarios). Look at QueryServicesClientImpl for examples.
+			throw new UnsupportedOperationException("This operation does not yet provide support for non-REST commands.");
+		}
+
+		if (result != null) {
+			return result.getItems();
+		} else {
+			return Collections.emptyList();
+		}
+	}
+	
 	@Override
 	public List<TaskInstance> findHumanTasksWithFilters(TaskQueryFilterSpec filterSpec, Integer page, Integer pageSize) {
