+++ kie-server-parent/kie-server-remote/kie-server-client/src/main/resources/META-INF/services/org.kie.server.client.helper.KieServicesClientBuilder
@@ -4,5 +4,5 @@ org.kie.server.client.helper.JBPMUIServicesClientBuilder
 org.kie.server.client.helper.OptaplannerServicesClientBuilder
 org.kie.server.client.helper.CaseServicesClientBuilder
 org.kie.server.client.helper.DMNServicesClientBuilder
-org.kie.server.client.helper.JBPMQueryServicesClientBuilder
+org.kie.server.client.helper.JBPMSearchServicesClientBuilder
